- minor fix for eventlog under NIX*
git-svn-id: svn://svn.zabbix.com/trunk@3731 f4e7fd02-d210-4e34-a1e0-d4093fb66cc2

+++ src/zabbix_agent/eventlog.h
@@ -22,7 +22,7 @@
 
 int	process_eventlog(
 		char *source,
-		int *lastlogsize, 
+		long *lastlogsize, 
 		char *timestamp, 
 		char *src, 
 		char *severity,
